Give Environment's own instance columns a single owner
`instances` is core's table, and `EnvironmentHandlerState` promised that "all
instance write operations are delegated to this shared persistence layer" while
five modules wrote it directly behind that promise. The writes are legitimate —
peak memory, CPU time, captured stderr and the crash-loop counters are what a
process did and whether a restart may relaunch it, and core reads none of them —
but scattering them across metrics, recovery_marks and runtime is why nobody
noticed the doc was false.

InstanceRepository owns those columns now. metrics keeps the OTLP vocabulary,
which is its actual job, and delegates the SQL; recovery_marks had one caller and
folds in entirely. The doc says what is true instead: core's view changes through
Persistence, with two named exceptions — these columns, and the pending-instance
insert inside claim_initial, which shares a transaction with the launch row so a
failure between the two cannot strand a pending instance no generation owns.

Three writers left where there were five, and the two that remain are the ones
with a reason.
